requirements.txt
transformers
torch
requests
Pillow
fastapi
uvicorn
streamlit
python-multipart
langchain 
langchain-core 
langchain-community 
pydantic 
duckduckgo-search
langchain_openai
langchain-groq
prometheus_client
accelerate
perft